Fix HumanReadable#fmt

+++ src/rustup-cli/download_tracker.rs
@@ -178,7 +178,7 @@ impl fmt::Display for HumanReadable {
             if sec.is_infinite() {
                 write!(f, "Unknown")
             } else if sec > 48. * 3600. {
-                let sec = self.0 as u64;
+                let sec = self.0;
                 let d = sec / (24. * 3600.);
                 let h = sec % (24. * 3600.);
                 let min = sec % 3600.;
@@ -186,14 +186,14 @@ impl fmt::Display for HumanReadable {
 
                 write!(f, "{:3} days {:2} h {:2} min {:2} s", d, h, min, sec) // XYZ days PQ h RS min TU s
             } else if sec > 6_000. {
-                let sec = self.0 as u64;
+                let sec = self.0;
                 let h = sec / 3600.;
                 let min = sec % 3600.;
                 let sec = sec % 60.;
 
                 write!(f, "{:3} h {:2} min {:2} s", h, min, sec) // XYZ h PQ min RS s
-            } else if sec > 100 {
-                let sec = self.0 as u64;
+            } else if sec > 100. {
+                let sec = self.0;
                 let min = sec / 60.;
                 let sec = sec % 60.;
